requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
FinanceDataReader>=0.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.18.0
dart-fss>=0.4.3
lxml>=4.9.0
